pyproject.toml
[project]
name = "llm_service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs==9.12.1",
    "fastapi[standard-no-fastapi-cloud-cli]==0.135.3",
    "ipython==9.11.0",
    "langchain==1.2.15",
    "langchain-community==0.4.1",
    "langchain-core==1.2.28",
    "langchain-openai==1.1.7",
    "langgraph==1.1.6",
    "langgraph-checkpoint==4.0.0",
    "langgraph-checkpoint-sqlite==3.0.3",
    "python-dotenv==1.2.1",
    "sqlite-vec==0.1.9",
]
